+++ R/facet-strata.R
@@ -9,14 +9,14 @@
 #'
 #' @examples
 #' library(ggplot2)
-#' ggplot(world_heights,
+#' ggplot(heights,
 #'        aes(x = year,
 #'            y = height_cm,
 #'            group = country)) +
 #'   geom_line() +
 #'   facet_strata()
 #' 
-#' ggplot(world_heights,
+#' ggplot(heights,
 #'        aes(x = year,
 #'            y = height_cm,
 #'            group = country)) +