+++ tests/testthat/test-facet-strata.R
@@ -2,7 +2,7 @@ context("facet-strata")
 
 library(ggplot2)
 set.seed(2019-07-23-1835)
-gg_facet_strata <- ggplot(world_heights,
+gg_facet_strata <- ggplot(heights,
                           aes(x = year,
                               y = height_cm,
                               group = country)) +